pyproject.toml
[tool.poetry]
name = "dria-searching-agent"
version = "0.1.0"
description = "Agentic search framewowk"
authors = ["andthattoo <omer@firstbatch.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
anthropic = "^0.25.7"
crewai = {extras = ["tools"], version = "^0.28.8"}
langchain = "^0.1.17"
unstructured = "^0.13.6"
playwright = "^1.43.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
langchain-anthropic = "^0.1.11"
bs4 = "^0.0.2"

[tool.poetry.group.dev.dependencies]
openai = "^1.25.1"
duckduckgo-search = "^5.3.0"

[tool.poetry.scripts]
app = "dria_searching_agent.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
